chore(tests): split #4898's big test into several test files
Also moved #4892 into the entity generator folder.

Moved schema creation in each file into a beforeAll(),
rather than at the start of every test.
