+++ tests/features/entity-generator/ScalarPropsForFks.mysql.test.ts
