+++ tests/features/entity-generator/TypesForScalarDecorators.mysql.test.ts
